fix:change type of id

+++ isa-project/src/main/java/rs/ac/uns/ftn/isaproject/service/users/UserAccountService.java
@@ -7,9 +7,9 @@ import rs.ac.uns.ftn.isaproject.model.users.User;
 import rs.ac.uns.ftn.isaproject.model.users.UserAccount;
 
 public interface UserAccountService {
-	UserAccount findById(int id) throws AccessDeniedException;
+	UserAccount findById(Long id) throws AccessDeniedException;
     UserAccount findByUsername(String username);
     void save(String username, String password, String role, boolean enabled, User user);
-    void updatePassword(int id,String oldPassword, String newPassword) throws BadRequestException;
+    void updatePassword(Long id,String oldPassword, String newPassword) throws BadRequestException;
 }
